requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
asyncpg==0.31.0
certifi==2026.2.25
charset-normalizer==3.4.5
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
dotenv==0.9.9
executing==2.2.1
flake8==7.3.0
fonttools==4.61.1
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipykernel==6.30.1
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
mccabe==0.7.0
mypy_extensions==1.1.0
narwhals==2.18.0
nest-asyncio==1.6.0
numpy==2.4.2
packaging==26.0
pandas==3.0.1
pandera==0.29.0
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.2
plotly==6.6.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.2.2
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2==2.9.11
pure_eval==0.2.3
pyarrow==23.0.1
pycodestyle==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.47
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.5.4
traitlets==5.14.3
typeguard==4.5.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
wcwidth==0.6.0
wordcloud==1.9.6
xgboost==3.2.0
